Added initial support for shutter calibration. At the moment a calibration file is simply produced by calibrate_shutters.py, and a comment in that file describes the future plans for how this will be produced. Calibrations are stored as attributes in a hdf5 file calibrations.h5.
When instantiating a shutter, a labscript user can add an argument: delay=(open_delay,close_delay), to specify the delays in the script, or can set delay='calibrated' to instruct labscript
to find the delays in the calibrations.h5 file. They are attributes to the group /shutters/shuttername, where shuttername is the name of the shutter object, given
to it as the first argument when the user instantiates it.

Also this commit makes devices attached to pulseblaster flags processed as being on the slow clock. There can be no constant clock rate ramps for such devices, since the pulseblaster
itself does not output buffered values when it gets a clock tick. Rather, a separate instruction is required every time an output changes state. The output produced is identical, but
internally labscript no longer expands the pulseblaster flags to their values at every time in the fast clock.

If this change hadn't been made, then we would have run into problems with the pulseblaster DDS outputs, once functionality for them is implemented.* It is nonsensical to ask one of them
to do a ramp on the fast clock, since every update requires a new pulseblaster instruction, and the fast clock is implemented with loops. Ramps will have to be done with repeated calls to
DDS.set_amp(t, value) etc. This can still be abstracted away to perform fast ramps (or ones with a non constant update rate), but there is a risk that these instructions will collide with
instructions on the devices attached to the flags, exceeding their maximum update rate. In this case labscript throws an error. Really the slow clock shouldn't be used that way, and if we
need pulseblaster or novatech boards to do ramps like that we should make sure nothing else on the pulseblaster flags needs to change state at the same time.

These are the sorts of things we should consider when deciding what thing to plug into what pulseblaster!