Analog inputs implemented with a class called AnalogIn, which is to be a child device of an NIBoard object. The list of analog inputs used is stored as an attribute to the group for the NI Card in the hdf5 file.
Acquisitions can be programmed with calls to AnalogIn.acquire(label,start_time,stop_time), with an optional scale factor and units string. These can be used later duting analysis.
The NIBoard class now takes an acquisition_rate argument when initialised, which is stored as an attribute as well. The LabVIEW control system will use this clock rate and the list
of analog inputs to program the boards to acquire data for the entirety of the experiment. The acquired data will have the same number of points for each input channel, and so can be
stored as a single array under the NI card's group in the hdf5 file.

A post-processing script will take this acquired data and split it up into separate measurements (corresponding to single calls to AnalogIn.acquire), using data for the start and stop
times of each call, stored as a table called ACQUISITIONS in the hdf5 file.

A couple of attributes changed name in this revision (to avoid ambiguities), so unless they were hardcoded in LabVIEW, things might break!