Device initialisation now requires only one name for each object, rather than two.
Previously, you had to both give the object a name as a string argument, and you had to specify a python variable name to reference it:

	name_1 = Device('name_2', *otherargs)

Now you only provide one name, as a string argument, and the initialisation will automatically place a reference to the device object in the
global namespace:

>>> Device('name_1', *otherargs)
>>> print name_1
<__main__.Device object at 0x7f0ca4a64e50>

Of course, the name must be a valid Python variable name -- else a ValueError will be raised.

This also means that you now have to explicitly give the full name of all the IntermediateDevices, such as NovatechDDS9ms, NIBoards, and also
PulseBlasters. Previously these objects made their own names from a fixed string, and just appended a number to it. We'll have to ensure that these
names match what LabVIEW expects -- so that it can match these devices to the actual hardware it programs.

This should be pretty easy to enforce, especially since I'm about to implement code generation for the device initialisation part of labscripts.

Stay tuned for that!