minor change to how params and device references are put into the global namespace. Now one can do:
        import labscript
        labscript.Device('device1', *otherargs)
        device1.do_something(t=7)
        print param1, param2
	labscript.stop(t=10)

rather than having to import * (which is generally frowned upon in production code), or reference the params
(which are the experimental run's 'global' arguments, as read from the hdf5 file) with labsript.param1 or similar.

I've left example.py as having 'import *' for the moment, which also imports * from pylab. I can't really see this
becoming an issue, but if it does we can just import labscript instead. Or maybe when I package labscript properly
I'll make an __all__ list that will only import the bits that the user is supposed to use. But it doesn't matter,
we're all grown ups here and don't have to have the code hidden from us to make sure we don't break things.

If it turns out people slip into the habit of breaking things (making names that conflict with what labscript uses etc),
then we'll work out what should be hidden based on that.