Now doesn't allow intereactive imports (since labscript relies on there being an actual script file). Also, doesn't create a calibrations file if it doesn't exist. Might ditch calibrations altogether depending on whether we end up using globals for that functionality.
System wide install should now work so long as labscript's parent directory is in the PYTHONPATH. This is probably how we should do things. On the master image I've added C:\pythonlib to the %PYTHONPATH% system variable,
and put labscript in C:\pythonlib\labscript. labscript should now be importable from anywhere on the system.