PulseBlaster DDSs now work. You can create a DDS object, declare its parent to be a pulseblaster, and set its amplitude, freq and phase as with the existing DDS objects.
The ranges are checked, and the number of registers are checked such that labscript will throw an error if an out of bounds value is requested or the allowed number of registers is exceeded. Note that I've hardcoded the register limits for the new firmware -- the old firmware has much lower limits and labscript will not throw an error if they are exceeded.

If anyone is actually going to use one of the old boards for its DDS output, I can put the error checking in easily enough, and maybe make the firmware version an argument to the creation of a PulseBlaster object. But I won't bother unless they're actually in use.