Fixed a bug where the first instruction was set to 0 if a value wasn't specified at t=0. It is now set to use self.default_value instead.
Started to implement changes to DDS where you can specify the digital gate channel for a Novatech DDS, and can change the enable state for the PulseBlaster DDS's

The Novatech DDS modifications should now work, but need testing. The Pulseblaster is not implemented fully yet.

None of this should affect any existing labscript files as everything should be backwards compatible with the previous syntax!