Added software limits for analog/DDS channels This only applies to labscript files, and is untested. Will test significantly once I've implemented the software limits in BLACS manual mode too
Renamed dds "en" channels in the connection table to "gate". SO if you have a DDS called xyz the gate will be in the connection table as "xyz_gate". This is to bring labscript in line with BLACS names, and should fix the saving and restoring of gate states issue when the connection table changes.

Note, your connection table for BLACS will need to be recompiled or else newly compiled experiments won't submit, even if the connection table python file has not changed (the internal h5 name for dds gates has changed and things will not match)