Fixed a bug.
Previously, if the ramp rate was so low that the time between instructions was greater than 55s (a rate of less than (1/110) Hz), the length between instructions would be out by half the remainder of the instruction length when divided by 55.

This is because we incremented the instruction counter j for the added LONG_DELAY instruction before making the END_LOOP instruction. We thus branched back to the LONG_DELAY instruction instead of the LOOP instruction defining the beginning of the loop. The length of time between the LOOP instruction and LONG_DELAY instruction was half the remainder of the instruction length when divided by 55 (where the instruction length is half the time between rising edges on the clock)