the exp_ramp function of AnalogQuantity did not return duration.
This bug will only have affected the spinorlab if they were calling:
t += channel.exp_ramp(....)
which I don't think they were (I don't think it was responsible for bug #511 as the ramp functions were not in use as the channels were not on the fast clock)

The bug was already fixed in the working copy of the dual_species lab, but not committed due to the function being modified further by Shaun.

The bug was fixed in the working copy of the spinor lab on Friday morning, but not committed.

It is now committed!