refs #534
The parameters for the camera object instantiation in labscript now include:
serial_number, SDK and effective_pixel_size and these are stored as attributes to each camera in the devices group.

*These parameters are required. See example.py for an example!*

SDK should be specified as a string
serial_number can be a hexadecimal number (0x1E) or a hexadecimal string (either "0x1E" or just "1E"). It is stored as a uint64 in the h5 file.
effective_pixel_size should be specified in units of meters as a single number [eg: 4.6e-6]. (we assume square pixels always, can add non square capability on request)