Fixed some bugs with the hardware triggering of a secondary PulseBlaster.
This mainly revolves around the staticDDS which inserts an instruction at t=0 to set it's output. However, due to the finite triggering time of the pulseblaster, even if you were to trigger it at t=0, no instructions can be issued on that device (or it's children) until t>=250ns. This causes problems when a Novatech is attached to a secondary pulseblaster.

The solution is to not store the StaticDDS (StaticAnalogQuantity) value in an instruction, but rather as a simple attribute of the class!