Fixed a bug which reared its head only under very specific circumstances. If a single element ramp (I know, why the hell would you do that?) was used, and immediately followed another ramp, then there was a strange interaction between the two.
The single element ramp would modify its function evaluation time to correct for zero order hold error. The problem was, it would modify an existing mutable data structure in doing so. The previous ramp would then see this modification, and if the two ramps had different sample rates, would get an incorrect time point for its own zero order hold correction. This resulted in a grossly incorrect final point for that first ramp, as its function was evaluated at some much later time.

The fix was to ensure that the correction for zero order hold error doesn't modify the shared time arrays in the case of a one-element ramp!

This was not a contrived example, this was actually discovered whilst in use. Sure a one-element ramp is unrealistic, but it was being used to essentially 'comment out' a ramp to be a single state change. It was more convenient to leave the ramp in and make it have only one point by modifying a runmanager global than to change the code.