Added /devices submodule for stuff that you might not want imported all the time.
Currently used only for ADWin implementation, which is currently in a working state. ADWin has a funny implementation with its own digital and analog output objects since they are not so interchangable with other devices. Ramps are implemented by programming the device with ramp parameters, which it evaluates itself.

This required an 'Instruction' object capable of holding the ramp parameters rather than simply the function for the ramp.