An attempted fix for the waits bug where absurdly short instructions were beign generated.
The fix was to round trigger times to the nearest 10ns like the rest of labscript times,
so that equality comparisons later were not subject to floating point rounding errors.

A better fix for this would be to actually subclass float myself such that all
arithmetic operations rounded, or all equality checks were in some tolerance so
we didn't have to manually pepper the code with these roundings. Or whatever.
Something to automate this.